chore: 오디오 관련 패키지 이동
