+++ src/main/java/com/jolupbisang/demo/application/audio/service/AudioService.java
@@ -12,12 +12,12 @@ import com.jolupbisang.demo.application.event.MeetingCompletedEvent;
 import com.jolupbisang.demo.application.event.MeetingStartingEvent;
 import com.jolupbisang.demo.application.event.whisper.WhisperEmbeddedEvent;
 import com.jolupbisang.demo.global.exception.CustomException;
+import com.jolupbisang.demo.infrastructure.audio.AudioProgressRepository;
+import com.jolupbisang.demo.infrastructure.audio.AudioRepository;
 import com.jolupbisang.demo.infrastructure.audio.EmbeddedVectorRepository;
 import com.jolupbisang.demo.infrastructure.audio.EmbeddingAudioRepository;
+import com.jolupbisang.demo.infrastructure.audio.client.WhisperClient;
 import com.jolupbisang.demo.infrastructure.aws.sfn.SfnClientUtil;
-import com.jolupbisang.demo.infrastructure.meeting.audio.AudioProgressRepository;
-import com.jolupbisang.demo.infrastructure.meeting.audio.AudioRepository;
-import com.jolupbisang.demo.infrastructure.meeting.client.WhisperClient;
 import com.jolupbisang.demo.infrastructure.meetingUser.MeetingUserRepository;
 import com.jolupbisang.demo.presentation.audio.dto.response.SocketResponseType;
 import lombok.RequiredArgsConstructor;
