+++ src/main/java/com/jolupbisang/demo/application/participationRate/ParticipationRateService.java
@@ -11,8 +11,8 @@ import com.jolupbisang.demo.domain.meeting.Meeting;
 import com.jolupbisang.demo.domain.participationRate.ParticipationRate;
 import com.jolupbisang.demo.domain.user.User;
 import com.jolupbisang.demo.global.exception.CustomException;
+import com.jolupbisang.demo.infrastructure.audio.client.dto.response.DiarizedResponse;
 import com.jolupbisang.demo.infrastructure.meeting.MeetingRepository;
-import com.jolupbisang.demo.infrastructure.meeting.client.dto.response.DiarizedResponse;
 import com.jolupbisang.demo.infrastructure.participationRate.ParticipationRateRepository;
 import com.jolupbisang.demo.infrastructure.sse.MeetingSseEventType;
 import com.jolupbisang.demo.infrastructure.sse.MeetingSseService;
