+++ src/main/java/com/jolupbisang/demo/application/segment/SegmentService.java
@@ -9,9 +9,9 @@ import com.jolupbisang.demo.application.segment.dto.SocketSegmentRes;
 import com.jolupbisang.demo.domain.meeting.Meeting;
 import com.jolupbisang.demo.domain.segment.Segment;
 import com.jolupbisang.demo.domain.user.User;
+import com.jolupbisang.demo.infrastructure.audio.AudioProgressRepository;
+import com.jolupbisang.demo.infrastructure.audio.client.dto.response.DiarizedResponse;
 import com.jolupbisang.demo.infrastructure.meeting.MeetingRepository;
-import com.jolupbisang.demo.infrastructure.meeting.audio.AudioProgressRepository;
-import com.jolupbisang.demo.infrastructure.meeting.client.dto.response.DiarizedResponse;
 import com.jolupbisang.demo.infrastructure.segment.SegmentRepository;
 import com.jolupbisang.demo.infrastructure.user.UserRepository;
 import com.jolupbisang.demo.presentation.audio.dto.response.SocketResponseType;
