+++ src/main/java/com/jolupbisang/demo/application/segment/dto/SocketSegmentRes.java
@@ -1,13 +1,14 @@
 package com.jolupbisang.demo.application.segment.dto;
 
-import com.jolupbisang.demo.infrastructure.meeting.client.dto.response.DiarizedResponse;
+import com.jolupbisang.demo.infrastructure.audio.client.dto.response.DiarizedResponse;
+
 import java.time.LocalDateTime;
 
 public record SocketSegmentRes(
-    LocalDateTime timestamp,
-    long userId,
-    int order,
-    String text
+        LocalDateTime timestamp,
+        long userId,
+        int order,
+        String text
 ) {
     public static SocketSegmentRes of(DiarizedResponse.Segment segmentData, LocalDateTime timestamp) {
         return new SocketSegmentRes(
