+++ src/main/java/com/jolupbisang/demo/application/summary/SummaryService.java
@@ -9,8 +9,8 @@ import com.jolupbisang.demo.application.summary.exception.SummaryErrorCode;
 import com.jolupbisang.demo.domain.meeting.Meeting;
 import com.jolupbisang.demo.domain.summary.Summary;
 import com.jolupbisang.demo.global.exception.CustomException;
+import com.jolupbisang.demo.infrastructure.audio.client.WhisperClient;
 import com.jolupbisang.demo.infrastructure.meeting.MeetingRepository;
-import com.jolupbisang.demo.infrastructure.meeting.client.WhisperClient;
 import com.jolupbisang.demo.infrastructure.sse.MeetingSseEventType;
 import com.jolupbisang.demo.infrastructure.sse.MeetingSseService;
 import com.jolupbisang.demo.infrastructure.summary.SummaryRepository;
