+++ src/main/java/com/jolupbisang/demo/infrastructure/audio/AudioRepository.java
@@ -1,4 +1,4 @@
-package com.jolupbisang.demo.infrastructure.meeting.audio;
+package com.jolupbisang.demo.infrastructure.audio;
 
 import com.jolupbisang.demo.application.audio.dto.AudioMeta;
 import org.springframework.stereotype.Repository;
@@ -12,6 +12,6 @@ public interface AudioRepository {
     String save(AudioMeta audioMeta, byte[] audioData) throws IOException;
 
     String findCompletedURLByMeetingIdAndUserId(long meetingId, long userId, Duration duration);
-    
+
     List<Long> findCompletedUserIdsByMeetingId(long meetingId);
 }
