+++ src/main/java/com/jolupbisang/demo/infrastructure/audio/client/dto/request/ReferenceRequest.java
@@ -1,4 +1,4 @@
-package com.jolupbisang.demo.infrastructure.meeting.client.dto.request;
+package com.jolupbisang.demo.infrastructure.audio.client.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
