+++ src/test/java/com/jolupbisang/demo/SilRokApplicationTests.java
@@ -3,7 +3,7 @@ package com.jolupbisang.demo;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class SilRokApplicationTests {
 
     @Test
